On avance bien la

+++ src/Controller/AdminUserController.php
@@ -3,8 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Form\AdminUserType;
-use App\Repository\UserRepository;
+use App\Form\AdminUserEditType;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -15,47 +14,68 @@ use Symfony\Component\Routing\Attribute\Route;
 #[Route('/admin/users', name: 'admin_users_')]
 class AdminUserController extends AbstractController
 {
-    #[Route('/', name: 'list')]
-    public function list(UserRepository $repo, Request $request, PaginatorInterface $paginator): Response
+    #[Route('/', name: 'index')]
+    public function index(Request $request, EntityManagerInterface $em, PaginatorInterface $paginator): Response
     {
-        $query = $repo->createQueryBuilder('u')
-            ->orderBy('u.id', 'DESC')
-            ->getQuery();
+        $query = $em->getRepository(User::class)
+            ->createQueryBuilder('u')
+            ->orderBy('u.id', 'ASC');
 
         $users = $paginator->paginate(
             $query,
             $request->query->getInt('page', 1),
-            20 // nombre de users par page
+            20
         );
 
         return $this->render('admin/users.html.twig', [
-            'users' => $users,
+            'users' => $users
         ]);
     }
 
     #[Route('/edit/{id}', name: 'edit')]
-    public function edit(User $user, Request $request, EntityManagerInterface $em): Response
-    {
-        $form = $this->createForm(AdminUserType::class, $user);
+    public function edit(
+        int $id,
+        Request $request,
+        EntityManagerInterface $em
+    ): Response {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        $user = $em->getRepository(User::class)->find($id);
+        if (!$user) {
+            throw $this->createNotFoundException('Utilisateur introuvable.');
+        }
+
+        $form = $this->createForm(AdminUserEditType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
-            return $this->redirectToRoute('admin_users_list');
+            $this->addFlash('success', 'Utilisateur mis à jour ✔️');
+
+            return $this->redirectToRoute('admin_users_index');
         }
 
         return $this->render('admin/user_edit.html.twig', [
-            'form' => $form,
-            'user' => $user
+            'form' => $form->createView(),
+            'user' => $user,
         ]);
     }
 
     #[Route('/delete/{id}', name: 'delete')]
-    public function delete(User $user, EntityManagerInterface $em): Response
+    public function delete(int $id, EntityManagerInterface $em): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        $user = $em->getRepository(User::class)->find($id);
+        if (!$user) {
+            throw $this->createNotFoundException('Utilisateur introuvable.');
+        }
+
         $em->remove($user);
         $em->flush();
 
-        return $this->redirectToRoute('admin_users_list');
+        $this->addFlash('success', 'Utilisateur supprimé ✔️');
+
+        return $this->redirectToRoute('admin_users_index');
     }
 }
